Pin `cuda-toolkit` version for dependencies (#43)
As pointed out by Vyas, `cuda-toolkit` is a meta package that does not
depend on `cuda-version`. Components of `cuda-toolkit` may depend on
`cuda-version`, which previously I relied on that dependency to get the
correct CTK components. Some unexpected CI
[**success**](https://github.com/NVIDIA/numbast/actions/runs/8897777445/job/24433476729#step:5:443)
has shown that the default compiler in CI environment mismatches that
defined for `cuda-version`, indicating a transitive dependency failure.
Therefore, this PR explicitly pins `cuda-toolkit` package version.