Create Conda Package for numba-extensions-bf16, and update build system accordingly. (#45)
Conda package for bfloat16 (and subsequently for numbast and
ast_canopy). This PR includes several temporary changes to the build
system:
- libastcanopy.so (previously mistakenly named liblibastcanopy.so) now
installed into `SKBUILD_PLATLIB`, not `conda/lib`. `ast_canopy` build
system is currently hand crafted so that this shared object is moved
into the right location. See `ast_canopy/build.sh` for detail. When
doing conda build, the build command is in `meta.yaml::build/script`
key. This reuses the `SP_DIR` path provided in the conda build
environment, instead of the dynamically determined
`stie.getsitepackages()` function in build.sh, which is more robust.

The artifacts of this PR:
- Establishes 2 new CI jobs that builds ast_canopy, numbast and bf16
conda package. Then creates a new conda environment in these jobs, add
the `conda_bld` folder as the new channel, and install bf16 conda
package. Run the tests in these environments. The 2 CI job are V100 and
A100 machines respectively. Hopefully they can replace our current test
system. Provided we setup a S3 bucket to upload the conda build
artifacts.
- The above verifies certain capability of our build system supporting
conda package delivery, but with some drawbacks due to manual
installation of `libastcanopy.so`. I'm working to renovate the
ast_canopy's build system so that `libastcanopy.so` installation is
managed by cmake, and `pylibastcanopy` discovers it automatically. This
makes sure that scikit-build-core and conda build all recognizes this
dependency.

TODO work:

- [ ] Refactor build system to reduce the manual wrangling needed in
this PR. Shown in #52
- [ ] Deprecate the python test CI job and use the conda build job
instead.

---------

Co-authored-by: Michael Yh Wang <wangm@nvidia.com>
Co-authored-by: isVoid <isVoid@users.noreply.github.com>