Improve Pytorch-Numba bfloat16 UX (#60)
This PR moves the ProxyTorch object into the bfloat16 package and made a
patch to Numba's `_LaunchConfiguration` object that creates a wrapper
for every pytorch tensor that is of bfloat16 data type. This makes torch
users using bfloat16 bindings with pytorch tensors no longer having to
modify their own code with the Numba kernel. The conversion happens
transparently.