Bypass Expanded Macros via a prefix allowlist (#97)
As a workaround feature for macro expansions, this PR adds an additional
feature to `ast_canopy`, allowing user to specify a list of prefixes as
bypasses for macro expanded functions. This is a WAR solution because it
assumes that the methods defined with anonymous filename are the same as
macro expanded functions, which is not necessarily true.

In the future, I'd love to see the implementation based on preprocessors
callbacks, using it to associate expanded function decls with
`spellingloc` - the location where it was expanded.

closes #98