Just build for one Python version in a build
This moves to using CI to parallelize the supported Python versions and
just building one package for each build call. As there are no tags use
the first commit to create `GIT_DESCRIBE_NUMBER` and get the short hash
that are used in custom build strings. This helps the solver know that
these builds are newer than those in the conda repositories.