Split project build and testing into separate jobs (#103)
This used #85 as a starting point based on current main. It separates
the build portion that happens on a CPU host lacking GPU, then uploads
the packages for download and testing on a host with a GPU. It adds
initial support for `sccache` and moves to using the CI matrix to
specify the various Python versions supported.

It was necessary to migrate to `conda build` from `conda mambabuild` as
`boa` was removed from the CI images. The recipes were altered to record
the short SHA and commits since first commit in the custom build string
along with the Python version to make it easier to verify what version
is installed.

---------

Co-authored-by: isVoid <isVoid@users.noreply.github.com>