Single Source of Truth of Version, allow querying in package, bump to 0.3.0 (#113)
Uses single source of truth for all versions inside the ast_canopy
package. This includes the C++ and the python part of the version. All
referring to `VERSION` file.

Notably, the c++ portion of ast_canopy does not define a new version but
instead rely on the version that defines for the python package to
maintain consistency.

The `VERSION` file now exists in the python package source tree, this is
to support editable install. For that, `meta.yaml` for all packages are
changed to locate the new VERSION address.

This PR also adds `__version__` attribute to the package so that they
can be queried.

---------

Co-authored-by: isVoid <isVoid@users.noreply.github.com>
Co-authored-by: Marcus D. Hanwell <mhanwell@nvidia.com>