Use Single Upload Job for CondaBuild (#124)
We have two jobs in publish workflow currently. Each build to the same
set of targets: `py3.10-3.13` x `aarch64/amd64`. As it turns out that
the `CPU` field has no effect to the build. If we run the publish job,
it builds the same targets twic. On upload, it runs into duplicate
artifact conflicts:
https://github.com/NVIDIA/numbast/actions/runs/14845380294/job/41677954654.

This PR removes one of the job and only use the amd64 instance to build
for all targets.

Co-authored-by: isVoid <isVoid@users.noreply.github.com>