Add CTPSD matcher, WAR nested `RecordDecl` (#126)
Due to an upstream ASTMatcher bug: 

https://github.com/llvm/llvm-project/issues/138782#issuecomment-2856591256

We cannot reliably skip matching nested `RecordDecl` inside a
`ClassTemplateParitialSpecializationDecl` node. And since invoking
`getTypeSize` on an uninstantiated record is generally a bad idea, (in
fact, it leads to infinite recursion currently). We use a WAR to skip
parsing any nested `RecordDecl` that exists inside a `CTPSD` node. It
works like this:

Make a CTPSD matcher that precedes all existing matcher. When we matched
a CTPSD node, use a separate `RecursiveASTVisitor` to traverse this node
and collect all `RecordDecl` from within. Memoize all `RecordDecl` ID in
the payload. When the matcher matches a `RecordDecl`, check if its ID is
within the memoization. If it is, skip parsing this node.

Proper CTPSD handling will be added in a future PR.

Co-authored-by: isVoid <isVoid@users.noreply.github.com>