Use function mangled name to deduplicate forward declared function decls (#142)
This PR makes use of the `mangled_name` attribute added in #139 to
remove duplicate declarations due to forward declarations in C++.

Currently, because we use a crude manner to track the number of bindings
created for each function, there can be duplicates if multiple forward
declarations are found from the code base. This is not ideal as it
bloats the binding file size. This PR make uses of mangled name, which
*uniquely distinguishes* functions that have the same function signature
in the same scope, effectively removing duplicate bindings created for
forward declarations.

---------

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>