Fix a Bug where GlobalDecl is constructed mistakenly when performing name mangles (#146)
In #145 we bumped the clangdev version to 19, which originally was
thought to be a solution to a segfault to the name mangling on a
`CXXConstructorDecl` node. However, later troubleshooting shows that the
issue is caused by the `GlobalDecl` node is constructed with the wrong
constructor. When creating a `GlobalDecl` from a ctor decl node, it
should explicitly specify the `CXXCtorType` in argument. In our most
common use case, we use `Ctor_Complete` for most cases.

This PR reverts the minimum clangdev pinning back to 18 as it turns out
to be irrelevant to the cause of crash.

---------

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>