Add Config Item to Allow Module Setup/Teardown Callback Functions (#148)
This PR adds `Module Callbacks` item, with two subcategory named `setup`
and `teardown` respectively. Each of which is a string that will be
filled into the `setup_callback` and `teardown_callback` attribute for
the `CUSource` object used to define the shim function. Either item is
optional, in case where the item is not defined, the corresponding
CUSource attribute will be left unmodified.

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>