Propagate CMAKE_PREFIX_PATH from environment variable properly to skbuild (#202)
In https://github.com/NVIDIA/numbast/pull/188 `build.sh` we captured
`CMAKE_PREFIX_PATH` *from environment* and pass it to astcanopy build
options. The argument specified in the docstring says that it is a
"semicolon separated list". This contradicts with cmake standard where
environment variable are "colon separated list", while cmake options
arguments are "semicolon separated list". Additionally
`ASTCANOPY_INSTALL_PATH` is concatenated to `CMAKE_PREFIX_PATH` with
semi-colon delimiter, further violating the convention.

This caused a bug where the downstream skbuild for pylibastcanopy picked
up the wrongly delimited list and failed to search the install path of
astcanopy.

This PR is a temporary fix for this issue. It implements two normalizer,
and uniformly converts the list into:
- colon separated list for environment variables
- semicolon separated list for cmake arguments

However, ideally we should rely on cmake loading the variables from
environment and not try to pass it via argument (we are not doing any
fancy parsing of the path list anyway).

---------

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>