CI updates for conda builds (#251)
Updates to the conda build recipe to use updated GCC 14 compilers, more
recent glibc 2.28, and to avoid C++ module scanning as it is not used.
Also moved the builder images forward. Ideally we would switch to
smaller images for build as we do not need CUDA etc for build. This
harmonizes with the compiler and glibc used in the wheels builds.