Use `TargetRegistry.lower_cast` in separate registry mode (#263)
In separate registry mode, new instances of typing and target registries
are created and added to CUDA's third party registry separately. This is
in contrast to directly registerying to `cudaimpl` registry.

We expose most of the `register` functions, except `lower_cast`, which
should also be shadowed by the custom target registry.

Co-authored-by: Michael Wang <isVoid@users.noreply.github.com>